backend/requirements.txt
Flask==3.1.2
Flask-SocketIO==5.5.1
Flask-CORS==5.0.0
google-generativeai==0.8.5
python-dotenv==1.2.1
python-socketio==5.15.0
python-engineio==4.12.3
